Add generator testcase for invalid product blocks config
