+++ test/unit_tests/cli/data/invalid_product_config1.yaml
@@ -1,0 +1,20 @@
+# Testcase for invalid config: multiple root product blocks
+config:
+  summary_forms: False
+name: invalidexample1
+type: InvalidExample4
+tag: INVALIDEXAMPLE1
+description: "Invalid Product example 1"
+product_blocks:
+  - name: block1
+    type: Block1
+    tag: BLOCK1
+    fields:
+      - name: num_val
+        type: int
+  - name: block2
+    type: Block2
+    tag: BLOCK2
+    fields:
+      - name: str_val
+        type: str
